+++ HISTORY.rst
@@ -9,8 +9,12 @@ Finish Functionality of whole work
 * change deal_domain to get response. set it into self.response
 * add grequests module.
 TODO friend links include redirect links from target site.
-TODO GLOBAL config. test if it can be implemented like tornado.options.
 TODO classified all the response interface into one.
 TODO two kind of settings. product env and dev env
 TODO add multiple processes
 TODO get_web_source self.url or self.domain should be careful
+
+### 2017-7-20
+* make sure GLOBAL config from other file can be changed
+* improve friend link accuracy by reuqests to domain url. probably it's a jump link
+# Does friend link means multiple of same domain links or ??